pyproject.toml
[tool.poetry]
name = "attestationproject"
version = "0.1.0"
description = ""
authors = ["Vladislav Zorin <zorin.3v6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
drf-yasg = "^1.21.10"
coverage = "^7.8.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"


[tool.poetry.group.dev.dependencies]
ipython = "^9.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
